Shuffled around some functions and renamed some files.
- changed the name of onsetStrat.[ch] -> transient.[ch]
- renamed the detectTransients function in pairTransientDetection.[ch] as SelectTransients
- renamed pairTransientDetection.[ch] -> selectTransient.[ch]
- moved the pairwiseTransientDetection function from selectTransient.[ch] to transient.[ch] and changed its name to DetectTransients
- changed the name of TransientDetectionStrategy in transient.[ch] to DetectTransientsFromResampled
- Moved rollSigma out of simpleDetFunc and improved error reporting
- Changed the name of simpleDetFunctionCalculation in simpleDetFunc.[ch] to CalcDetFunc
- renamed simpleDetFunc.[ch] -> calcDetFunc.[ch]
