+++ src/errors.def
@@ -57,11 +57,20 @@ ERROR_ENTRY(ME_BAD_HOPSIZE, -204,
   "integral multiple of 4")
 
 
-ERROR_ENTRY(ME_BAD_KERNEL_FIT, -240,
+
+ERROR_ENTRY(ME_BAD_NUM_FILTER_CHANNELS, -211,
+  "The specified number of channels was 0 or larger than expected.")
+
+ERROR_ENTRY(ME_BAD_DETECTION_FUNCTION_LEN, -212,
+  "The length of the detection function buffer is incorrect")
+
+
+
+ERROR_ENTRY(ME_BAD_KERNEL_FIT, -241,
   "There was an issue while trying to identify the best-fit kernel.")
-ERROR_ENTRY(ME_NO_TRANSIENTS, -241,
+ERROR_ENTRY(ME_NO_TRANSIENTS, -242,
   "No tranisents could be identified.")
-ERROR_ENTRY(ME_ALL_NULL_DETFUNC, -242,
+ERROR_ENTRY(ME_ALL_NULL_DETFUNC, -243,
   "All entries in the detection function were zero.")
