+++ src/transient/rollSigma.c
@@ -1,13 +1,5 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <math.h>
-#include <time.h>
-#include "filterBank.h"
-#include "calcSummedLagCorrentrograms.h"
-#include "../errors.h"
-#include "../utils.h" // AlignedAlloc
-
-#define MAX_CHANNELS 128
+#include <stdlib.h> // malloc, free
+#include <math.h> //sqrtf
 
 /* Concept is taken from python Pandas package*/
 
@@ -158,7 +150,6 @@ void rollSigma(int startIndex, int interval, float scaleFactor,
 		 * never removed */
 		if (i == 0){
 			for (j=winStart;j<winStop;j++){
-				//printf("%f\n",buffer[j]);
 				add_var((double)buffer[j], &nobs, &mean_x,
 					&ssqdm_x);
 			}
@@ -223,140 +214,4 @@ void rollSigma(int startIndex, int interval, float scaleFactor,
 		}
 	}
 	windowIndexerDestroy(wInd);
-}
-
-int simpleComputePSM(int numChannels, const float * restrict data,
-		     float * restrict buffer,
-		     const float * restrict centralFreq,
-		     int sampleRate, int dataLength, int startIndex,
-		     int interval, float scaleFactor, int sigWindowSize,
-		     int numWindows, float * restrict sigmas,
-		     int correntropyWinSize,
-		     float * restrict pooledSummaryMatrix)
-{
-	float loop_time = 0.0f;
-	float correntropy_time = 0.0f;
-	for (int i = 0;i<numChannels;i++){
-		clock_t entry_c = clock();
-
-		int tmp_rslt;
-
-		tmp_rslt = sosGammatone(data, buffer, centralFreq[i],
-					sampleRate, dataLength);
-		if (tmp_rslt != ME_SUCCESS){
-			return tmp_rslt;
-		}
-
-		/* compute the sigma values */
-		rollSigma(startIndex, interval, scaleFactor, sigWindowSize,
-			  dataLength, numWindows, buffer, sigmas);
-
-		clock_t precorrentrogram_c = clock();
-		/* compute the pooledSummaryMatrixValues */
-		tmp_rslt = CalcSummedLagCorrentrograms(
-			buffer, sigmas, (size_t) correntropyWinSize,
-			(size_t) correntropyWinSize, (size_t) interval,
-			(size_t) numWindows, pooledSummaryMatrix);
-		if (tmp_rslt){
-			return tmp_rslt;
-		}
-
-		clock_t exit_c = clock();
-		loop_time += ((float)(exit_c-entry_c))/CLOCKS_PER_SEC;
-		correntropy_time +=
-			((float)(exit_c-precorrentrogram_c))/CLOCKS_PER_SEC;
-	}
-	printf("  average correntropy time (seconds): %e\n",
-	       correntropy_time / numChannels);
-	printf("  average total loop time (seconds):  %e\n",
-	       loop_time / numChannels);
-	return 0;
-}
-
-int computeNumWindows(int dataLength, int correntropyWinSize, int interval)
-{
-	int numWindows = (int)ceil((dataLength - correntropyWinSize)/
-				   (float)interval) + 1;
-	return numWindows;
-}
-
-int computeDetFunctionLength(int dataLength, int correntropyWinSize,
-			     int interval)
-{
-	return computeNumWindows(dataLength, correntropyWinSize, interval) - 1;
-}
-
-int simpleDetFunctionCalculation(int correntropyWinSize, int interval,
-				 float scaleFactor, int sigWindowSize,
-				 int numChannels, float minFreq, float maxFreq,
-				 int sampleRate, int dataLength, float* data,
-				 int detFunctionLength, float* detFunction)
-{
-
-	// TODO change the type of all length/index variables to size_t
-	// TODO Refactor so that this function doesn't allocate memory from the
-	//      heap (i.e. use a fixed buffer allocated on the stack)
-	// TODO Add meaningful error codes
-
-	if (numChannels > MAX_CHANNELS) {
-		return -1;
-	}
-	float centralFreq[MAX_CHANNELS];
-	centralFreqMapper(numChannels, minFreq, maxFreq, centralFreq);
-
-	// check correntrogram properties. Doing this now ensures that we
-	// compute a buffer size that is an integral multiple of 4
-	int arg_check = CheckCorrentrogramsProp((size_t) correntropyWinSize,
-						(size_t) correntropyWinSize,
-						(size_t) interval);
-	if (arg_check != 0){
-		return arg_check;
-	}
-
-	int numWindows = computeNumWindows(dataLength, correntropyWinSize,
-					   interval);
-	printf("datalen: %d, numWindows: %d\n", dataLength, numWindows);
-	if (detFunctionLength != (numWindows-1)){
-		return -1;
-	}
-
-	// For performance reasons, CalcSummedLagCorrentrograms doesn't process
-	// filterred data with arbitrary lengths. It requires that the buffer
-	// holds the following number of entries:
-	int bufferLength = (numWindows-1)*interval + 2*correntropyWinSize;
-	printf("bufferLength: %d\n",bufferLength);
-	float * buffer = AlignedAlloc(16, sizeof(float)*bufferLength);
-	// We accomodate this by simply setting all elements with indices >=
-	// dataLength to zero.
-	// TODO: Improve handling of alternative lengths. It would be more
-	//       correct to pad the end of the input signal with 0s
-	for (int i = dataLength; i < bufferLength; i++){
-		buffer[i] = 0;
-	}
-
-	// Prepare buffer for pooledSummaryMatrix
-	float * pooledSummaryMatrix = malloc(sizeof(float)*numWindows);
-	for (int i = 0; i < numWindows; i++){
-		pooledSummaryMatrix[i] = 0;
-	}
-
-	float * sigmas = malloc(sizeof(float)*numWindows);
-	int startIndex = correntropyWinSize/2;
-
-	int result = simpleComputePSM(numChannels, data, buffer, centralFreq,
-				      sampleRate, dataLength, startIndex,
-				      interval, scaleFactor, sigWindowSize,
-				      numWindows, sigmas, correntropyWinSize,
-				      pooledSummaryMatrix);
-	if (result == 0){
-		for (int i = 0; i < detFunctionLength; i++){
-			detFunction[i] = (pooledSummaryMatrix[i+1]
-					  - pooledSummaryMatrix[i]);
-		}
-	}
-
-	free(sigmas);
-	free(buffer);
-	free(pooledSummaryMatrix);
-	return result;
 }
